0. Print list

+++ 0x12-singly_linked_lists/0-main.c
@@ -27,15 +27,14 @@ int main(void)
     new->next = head;
     head = new;
     n = print_list(head);
-    printf("-> %lu elements\n", (unsigned long)n);
+    printf("-> %lu elements\n", n);
 
     printf("\n");
     free(new->str);
     new->str = NULL;
     n = print_list(head);
-    printf("-> %lu elements\n", (unsigned long)n); 
+    printf("-> %lu elements\n", n);
 
     free(new);
     return (0);
 }
-
